Fix enable_captcha option
Prior to this commit, when subscribing to a person's updates,
captcha box is opened. Given that it is an experimental feature
in people finder project and our team wanted to disable it,
this patch takes care of disabling it.
Also, captcha was tightly coupled with subcription.
This patch decouples captcha and subsription functionality.