Update unit_tests.py to not change pytest option parsing.
unit_tests.py previously changed an internal pytest option parsing
function, as a way of using the TESTS_DIR directory by default. pytest
now has a way to be called programmatically with arguments:
https://docs.pytest.org/en/latest/usage.html#calling-pytest-from-python-code
So the unit tests script can just check if any of the command-line
arguments are a test directory or something else to test (vs. just
options like --tb=short) and, if none of them are, add TESTS_DIR as a
default. Then that list of arguments can be passed to pytest.main.