Translation files for development.
I don't have a pipeline for generating these files for real yet, but
it's useful to have a couple language files just to be sure the
language/translation code is working. The Spanish file just has two
messages as it's just proof-of-concept (for anything else it'll use the
default message). For English, it doesn't really need anything in the
file since it's the default language anyway (at some point I'd like to
explore whether it's possible to avoid including the default messages in
the minified JS, but that doesn't seem like an immediate requirement).