Add some absolute_import imports.
For const.py, its unit test file, and amp_start.py. These files don't appear to
need any changes to be compatible with Python 3. Adding absolute_import will
make them use Python 3's default importing behavior (so we don't accidentally
mess it up later). Plus, python-modernize adds it to all files, and we want to
be able to use it for enforcement later (so anything it adds, we kinda have to
add).