Upgrade lxml to version 3.7.3.
This isn't the latest version of lxml, but it is the latest version that App
Engine supports (we can't app/vendors it because it's a wrapper around C
libraries). It's listed as supporting up to Python 3.5, which hopefully means
it'll work on 3.7 too (version 2.3 only officially supports up to Python 3.2).