Upgrade to WebOB 1.2.3. (#580)
The motivation for this is to keep us from breaking if/when App Engine
upgrades its Python 2.7 runtime to 2.7.15 (see issue #575). I'm
upgrading webapp2 to 2.5.2 too, because webapp2 depends heavily on WebOB
so it just seems better to run versions from similar times.

I'm also dropping support for character sets other than UTF-8, at least
for now. The reason for this is that WebOB made a change to how it
handles that which there's no good way to handle with any verison of
webapp2 (further details about this are also on issue #575). I'd like to
revisit this once we migrate to Django (which should make it
considerably easier to handle other charsets), but we've talked this
over with the team in Japan and it's not worth the work to support other
charsets with webapp2.