Python 3 testing. (#582)
This sets up Python 3 testing for both local development and Travis CI, for the (currently quite limited) set of tests that we expect to pass with Python 3 (listed in tests/unit_tests.py).
As part of doing this, I changed the way our Travis tests run by taking them out of Docker and running on Travis directly. As far as I know there wasn't a particular reason to run them in a Docker container; I guess once you have a Docker setup it might just be the easiest path to Travis tests, but it's kinda limiting if you need to run tests in multiple environments. We'll keep the Docker setup around though (it's useful for other things).
Also, minor change: we no longer specify the version of pytest to install, so pip will just get the latest version. We might be able to do the same with Pillow later, but I haven't looked into it (App Engine might be using an older version, so maybe we should be too).
Fixes issue #508.