Run Travis with both Python 2.7.12 and a recent 2.7.
We recently discovered that we would break with Python 2.7.15 (see issue
575). We'd been ok because App Engine uses Python 2.7.12 for its Python
2.7 runtime. However, it seems possible that they could upgrade to a
more recent version at any time, without giving us a heads-up (since no
one generally expects a patch version to be a breaking change). So, I
want to run our Travis tests with both Python 2.7.12 (because it's what
we currently get in prod) and a recent version (because we might get
something more recent in prod at any time).