Stop using jQuery.
We're not going to use jQuery with the React UI, and it seems like
overkill to use it for a lite mode, and I'd rather not load it if we
don't have to (especially since we have to avoid using a CDN to support
zero-rating). We're also not really using it much currently anyway: I
started seeing what it would take to upgrade to jQuery 3, and it turned
out to be easier to just rewrite the jQuery parts to use vanilla JS.

This does mean losing the sliding animations for open sections on
create/add_note pages (e.g., the "more information to identify the
person" section), but that seems ok, since I'm not sure we'd really want
JS animations in lite mode anyway. (If we really wanted to keep them
though, we could rewrite them in vanilla JS as well.)