Fuzzify ages.
For data we collect from users, fuzzify ages immediately; we don't
ever want to display or share the exact age. For data exchanged
between partners via our API, we want to be an honest broker so we
don't modify their data, but we do fuzzify it when it's displayed on
our site.