Don't include test translate_api_key in setup_pf.py.
I think this was intended to help make it easier for developers to get
started, but it means that we're putting an API key out in the world.
You don't need a key to develop anything that isn't directly related to
use of the Translate API (the Translation features get disabled
automatically if the config doesn't include a Translate key), so we can
just drop it from the setup script.