Drop standalone param.
This seems to only ever be read, never actually used. It's unclear what
its purpose was (it was part of the original PF code checkin), but its
only effect is to disable the map (maybe it did other things in the past
and what we see now is just leftover code that didn't get cleaned up).