Set up Django app to run alongside the webapp2 app.
This is not meant for submission as-is. Our prod instance serves from a
subdirectory (google.org/personfinder), which makes anything related to
URLs challenging to test locally. If "personfinder" is considered the
script prefix, I think we should be able to handle URLs the same locally
and in prod, as long as we stick to using request.path_info (instead of
request.path) when we need to access URLs directly, as described here:
https://docs.djangoproject.com/en/1.11/ref/request-response/#django.http.HttpRequest.path_info

However, I'm not 100% sure about that, and I don't see a way to verify
it without trying it out in prod. This change contains configuration for
a Django app that serves a couple test pages that display information
about various variables and functions from the Django HttpRequest
object, so we can see for sure how it works. My plan is to try this in
prod, but not actually merge this PR into the master branch. Later I'll
get a real page running with Django and send that out as a separate PR
(and then I'll start gradually moving more pages to Django until we're
done with webapp2 altogether).