Drop the AdminEnv class.
This is only used in one place, and each field is only referenced once,
so I don't think we need it anymore. It's kinda a problem because it
causes a shared template to depend directly on the request object. This
is particularly an issue now because it's an obstacle to the incremental
migration from webapp2 to Django (which has different request objects).