Set up initial Django app. (#620)
* Set up Django app to run alongside the webapp2 app.

This is not meant for submission as-is. Our prod instance serves from a
subdirectory (google.org/personfinder), which makes anything related to
URLs challenging to test locally. If "personfinder" is considered the
script prefix, I think we should be able to handle URLs the same locally
and in prod, as long as we stick to using request.path_info (instead of
request.path) when we need to access URLs directly, as described here:
https://docs.djangoproject.com/en/1.11/ref/request-response/#django.http.HttpRequest.path_info

However, I'm not 100% sure about that, and I don't see a way to verify
it without trying it out in prod. This change contains configuration for
a Django app that serves a couple test pages that display information
about various variables and functions from the Django HttpRequest
object, so we can see for sure how it works. My plan is to try this in
prod, but not actually merge this PR into the master branch. Later I'll
get a real page running with Django and send that out as a separate PR
(and then I'll start gradually moving more pages to Django until we're
done with webapp2 altogether).

* Clarify comment.

* Try setting FORCE_SCRIPT_NAME.

* Revert "Try setting FORCE_SCRIPT_NAME."

This reverts commit 1484853986a674f1b12a6278a6896f86f91a7c99.

* Move app/vendors setup to wsgi.py.

* most of admin/statistics on Django

* Fix import style in urls.py.

* Set action based on class, and pylint views.base.

* pylint views.admin.base and .statistics.

* Remove webapp2 admin statistics handler.

* Run yapf over new files.

* more pylinting

* Start tests for Django admin statistics view.

* Fix lang setting and finish tests.

* Fix non-admin user case.

* Really fix it.

* first round of changes from review

* second batch of changes from review

* add some copyright headers

* Use wildcard for allowed hosts.

* Add TODO about ALLOWED_HOSTS.

* changes to urls.py from review

* Simplify path-building function.

* Add comment for wsgi.application name.

* changes from review