Tests for access restriction of (Django-served) admin pages. (#623)
This sets up tests to check that admin pages in the Django server are access-controlled. It doesn't test admin pages for the non-Django server (which is currently all but one of them), but we're going to move everything to Django over time, so it doesn't seem worth it to set up a way to test stuff on the old web framework.

My goals were to:
- Ensure that nobody can forget to add tests to check that an admin page is access-restricted.
- Make it readable.
- Make it easy to add those tests.

I think I got 1 and 3; I'm not sure this is going to continue to be readable once I get into similar tests for XSRF checks, but hopefully it'll be ok.