Drop YAPF from Travis tests for now. (#632)
Although YAPF docs suggest using it as part of a CI workflow, we've run
into a risk associated with that: YAPF might change its implementation,
and if its output changes your CI will break. This may not sound like a
significant risk, but we set up YAPF for two files on Thursday and it
broke our Travis tests on Sunday.

This takes YAPF out of our Travis setup to unbreak things. I've left the
YAPF option in the tools/lint script for now, since I'm not sure what
the new plan is going to be and we might still end up using it (even if
it's not part of our Travis setup).