Allow non-integer timezone offsets. (#665)
The field in the admin page is marked as a number; by default, the
step size is considered to be 1, and this can cause browsers to
prohibit non-integer values. Setting the step to "any" allows any
numeric value.

We could set the step to 0.25 (anything more granular than that is
probably errorneous input), but that would cause the stepper buttons
to increment/decrement by 0.25, which could be annoying. Setting step
to "any" allows any numeric value but leaves the stepper buttons at
the default step of 1 (at least in Firefox and Chrome):
https://stackoverflow.com/questions/19011861/is-there-a-float-input-type-in-html5