Adding boilerplate notice to files where missing. (#679)
* Adding boilerplate notice to files where missing.

I see that many files are missing the boilerplate notice
which are need as per Apache 2.0 License. As standed in
appendix of apache 2.0 License:
APPENDIX: How to apply the Apache License to your work.

To apply the Apache License to your work, attach the following
boilerplate notice, with the fields enclosed by brackets "[]"
replaced with your own identifying information. (Don't include
the brackets!)  The text should be enclosed in the appropriate
comment syntax for the file format. We also recommend that
file or class name and description of purpose be included on the
same "printed page" as the copyright notice for easier
identification within third-party archives.

Copyright [yyyy] [name of copyright owner]

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

     http://www.apache.org/licenses/LICENSE-2.0
Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.

Signed-off-by: kshithijiyer <kshithij.ki@gmail.com>

* Fixing year and other issues caught by CI.

Signed-off-by: kshithijiyer <kshithij.ki@gmail.com>

* Removing boilerplate notice from recaptcha files.

Signed-off-by: kshithijiyer <kshithij.ki@gmail.com>

* Removing python2.7 shebang.

Signed-off-by: kshithijiyer <kshithij.ki@gmail.com>

* Removing boilerplate notice from docker shell files.

Signed-off-by: kshithijiyer <kshithij.ki@gmail.com>