Add npm audit option to tools/ui. (#712)
I don't think it makes sense to add this to the automated tests on
GitHub, because it can fail for reasons unrelated to the code being
changed. However, I do want to add it to the pre-release tests once the
React UI is used in production, and this makes that a little more
convenient.