Fix photo upload handling in Django.
Django handles uploaded files differently than webapp2.
Also refactors create.py to import the photo module. This is what's
recommended by the style guide, and this is a convenient (by convenient,
I mean mandatory) time to do that because the mocking library doesn't
seem to handle the situation well otherwise.